chore: update version of kadent_engine

+++ src/storage/project/trait_impl/graph.rs
@@ -12,15 +12,15 @@ use std::io::{Cursor, Read};
 impl AsBytes for Graph {
     fn as_bytes(&self, bytes: &mut Vec<u8>) {
         // Write input_id and output_id
-        bytes.extend((self.get_input_id().0 as u64).to_le_bytes());
-        bytes.extend((self.get_output_id().0 as u64).to_le_bytes());
+        bytes.extend(self.get_input_id().0.to_le_bytes());
+        bytes.extend(self.get_output_id().0.to_le_bytes());
 
         // Write each node
         let mut nodes_bytes = Vec::new();
         for (node_id, node) in self.get_node_map() {
             let mut node_bytes = Vec::new();
             node.as_bytes(&mut node_bytes);
-            nodes_bytes.extend((node_id.0 as u64).to_le_bytes());
+            nodes_bytes.extend(node_id.0.to_le_bytes());
             nodes_bytes.extend((node_bytes.len() as u64).to_le_bytes());
             nodes_bytes.extend(node_bytes);
         }
@@ -31,9 +31,9 @@ impl AsBytes for Graph {
         let edges = self.get_edges();
         bytes.extend((edges.len() as u64).to_le_bytes());
         for (from_id, out_idx, to_id, in_idx) in edges {
-            bytes.extend((from_id.0 as u64).to_le_bytes());
+            bytes.extend(from_id.0.to_le_bytes());
             bytes.extend((*out_idx as u64).to_le_bytes());
-            bytes.extend((to_id.0 as u64).to_le_bytes());
+            bytes.extend(to_id.0.to_le_bytes());
             bytes.extend((*in_idx as u64).to_le_bytes());
         }
     }
@@ -47,9 +47,9 @@ impl FromBytes for Graph {
 
         // Read input_id and output_id
         cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
-        let input_id = NodeID(u64::from_le_bytes(buf) as usize);
+        let input_id = NodeID(u64::from_le_bytes(buf));
         cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
-        let output_id = NodeID(u64::from_le_bytes(buf) as usize);
+        let output_id = NodeID(u64::from_le_bytes(buf));
 
         // Read nodes
         cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
@@ -61,7 +61,7 @@ impl FromBytes for Graph {
             nodes_cursor
                 .read_exact(&mut buf)
                 .with_ctx(ParseContext::Graph)?;
-            let node_id = NodeID(u64::from_le_bytes(buf) as usize);
+            let node_id = NodeID(u64::from_le_bytes(buf));
             nodes_cursor
                 .read_exact(&mut buf)
                 .with_ctx(ParseContext::Graph)?;
@@ -80,11 +80,11 @@ impl FromBytes for Graph {
         let edge_count = u64::from_le_bytes(buf) as usize;
         for _ in 0..edge_count {
             cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
-            let from_id = NodeID(u64::from_le_bytes(buf) as usize);
+            let from_id = NodeID(u64::from_le_bytes(buf));
             cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
             let out_idx = u64::from_le_bytes(buf) as usize;
             cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
-            let to_id = NodeID(u64::from_le_bytes(buf) as usize);
+            let to_id = NodeID(u64::from_le_bytes(buf));
             cursor.read_exact(&mut buf).with_ctx(ParseContext::Graph)?;
             let in_idx = u64::from_le_bytes(buf) as usize;
             graph.add_edge_unchecked((from_id, out_idx, to_id, in_idx));
